Added support for changing uid/gid before running the subcommand. - Also updated several comments, including bug descriptions. - Cleaned up type conversion errors (silenced compiler warnings). - Polished up some error handling.
The new code to handle user and group could be split out into a separate
module.  The general design is relatively sensible, but it is more
difficult to see because of the way the types and functions are
scattered throughout the source at present.

In object oriented terms:
  * The attributes are the new and old uid and gid.
  * The initialization of the attributes takes place at startup, when the
    command-line parameters are parsed.
  * If no new uid/gid are specified, there is no object (a global pointer
    is set to NULL).
  * There are methods for setting the [uid,gid] pair, and for resetting
    them afterwards.

Right now, we set the effective [uid,gid] before creating the subprocess
to run the shell command, then reset them afterwards.  It might make
more sense to just switch globally to the specified [uid,gid] at the end
of program startup.  This should be considered.